backend/requirements.txt
asgiref==3.10.0
certifi==2025.11.12
charset-normalizer==3.4.4
dj-database-url==3.0.1
Django==5.2.7
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
docopt==0.6.2
idna==3.11
pipreqs==0.4.13
psycopg2-binary==2.9.11
PyJWT==2.10.1
python-dotenv==1.2.1
requests==2.32.5
sqlparse==0.5.3
urllib3==2.5.0
yarg==0.1.10
